execve error handling in progress

+++ incs/shell.h
@@ -6,7 +6,7 @@
 /*   By: rucosta <rucosta@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/25 00:29:36 by fgameiro          #+#    #+#             */
-/*   Updated: 2026/03/31 19:52:20 by rucosta          ###   ########.fr       */
+/*   Updated: 2026/04/07 03:17:04 by rucosta          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct s_shell
 {
     t_env   *env;
     t_cmd   *cmds;
-    int     exit_status;
+    short     exit_status;
     int     cmd_count;
 }   t_shell;
